adding core to better segment code
